+++ src/Ext/WarheadType/Body.cpp
@@ -121,6 +121,11 @@ void WarheadTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->AnimList_ShowOnZeroDamage.Read(exINI, pSection, "AnimList.ShowOnZeroDamage");
 	this->DecloakDamagedTargets.Read(exINI, pSection, "DecloakDamagedTargets");
 	this->ShakeIsLocal.Read(exINI, pSection, "ShakeIsLocal");
+	this->ChangeOwner.Read(exINI, pSection, "IsOwnerChanger");
+	this->ChangeOwner_MindControl.Read(exINI, pSection, "IsOwnerChanger.SetAsMindControl");
+	this->ChangeOwner_Anim.Read(exINI, pSection, "IsOwnerChanger.ControlAnim");
+	this->ChangeOwner_Threshold.Read(exINI, pSection, "IsOwnerChanger.HealthThreshold");
+	this->ChangeOwner_AffectElites.Read(exINI, pSection, "IsOwnerChanger.AffectElites");
 
 	// Crits
 	this->Crit_Chance.Read(exINI, pSection, "Crit.Chance");
@@ -268,6 +273,7 @@ void WarheadTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 		|| this->Convert_Pairs.size() > 0
 		|| this->InflictLocomotor
 		|| this->RemoveInflictedLocomotor
+		|| this->ChangeOwner
 	);
 }
 
@@ -290,6 +296,11 @@ void WarheadTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->AnimList_ShowOnZeroDamage)
 		.Process(this->DecloakDamagedTargets)
 		.Process(this->ShakeIsLocal)
+		.Process(this->ChangeOwner)
+		.Process(this->ChangeOwner_MindControl)
+		.Process(this->ChangeOwner_Anim)
+		.Process(this->ChangeOwner_Threshold)
+		.Process(this->ChangeOwner_AffectElites)
 
 		.Process(this->Crit_Chance)
 		.Process(this->Crit_ApplyChancePerTarget)