+++ src/Ext/WarheadType/Detonate.cpp
@@ -135,6 +135,9 @@ void WarheadTypeExt::ExtData::DetonateOnOneUnit(HouseClass* pHouse, TechnoClass*
 	if (this->RemoveMindControl)
 		this->ApplyRemoveMindControl(pHouse, pTarget);
 
+	if (this->ChangeOwner)
+		this->ApplyOwnerChange(pHouse, pTarget);
+
 	if (this->Crit_Chance && (!this->Crit_SuppressWhenIntercepted || !bulletWasIntercepted))
 		this->ApplyCrit(pHouse, pTarget, pOwner);
 
@@ -244,6 +247,61 @@ void WarheadTypeExt::ExtData::ApplyRemoveDisguiseToInf(HouseClass* pHouse, Techn
 	{
 		if (pInf->IsDisguised())
 			pInf->Disguised = false;
+	}
+}
+
+void WarheadTypeExt::ExtData::ApplyOwnerChange(HouseClass* pHouse, TechnoClass* pTarget)
+{
+	const auto pExt = TechnoExt::ExtMap.Find(pTarget);
+	auto armorType = pTarget->GetTechnoType()->Armor;
+
+	if (pExt->Shield && pExt->Shield->IsActive() && !pExt->Shield->CanBePenetrated(this->OwnerObject()))
+		armorType = pExt->Shield->GetArmorType();
+
+	const auto pOwnerAnimType = this->ChangeOwner_Anim.Get(nullptr);
+	const double ownerChangeHealthThreshold = this->ChangeOwner_Threshold;
+	const bool isMindControl = this->ChangeOwner_MindControl;
+	const bool doesAffectElites = this->ChangeOwner_AffectElites;
+	const bool targetImmuneToPsionics = pTarget->GetTechnoType()->ImmuneToPsionics;
+
+	bool pImmune = (isMindControl && targetImmuneToPsionics);
+	if(pTarget->GetHealthPercentage() > ownerChangeHealthThreshold)
+		pImmune = true;
+
+	if(!doesAffectElites && (pTarget->Veterancy.IsElite()))
+		pImmune = true;
+
+	if (!pImmune)
+	{
+		if ((GeneralUtils::GetWarheadVersusArmor(this->OwnerObject(), armorType) != 0.0) && (!pTarget->IsMindControlled()))
+		{
+			pTarget->SetOwningHouse(pHouse, true);
+			if (pOwnerAnimType)
+			{
+				CoordStruct OwningAnimLocation = pTarget->Location;
+				if (isMindControl)
+					OwningAnimLocation.Z += pTarget->GetTechnoType()->MindControlRingOffset;
+				if (auto const pOwnerAnim = GameCreate<AnimClass>(pOwnerAnimType, OwningAnimLocation))
+				{
+					pOwnerAnim->Owner = pHouse;
+					if (isMindControl)
+					{
+						pTarget->MindControlRingAnim = pOwnerAnim;
+						pTarget->MindControlledByAUnit = true;
+					}
+					else
+					{
+						pOwnerAnim->SetOwnerObject(pTarget);
+						pOwnerAnim->Owner = pHouse;
+					}
+				}
+			}
+			else
+			{
+				if (isMindControl)
+					pTarget->MindControlledByAUnit = true;
+			}
+		}
 	}
 }